feat: add release script (#62)
## What I'm changing

- Add a script to automate tag creation

## How I did it

When I was cutting the release for **stapi-pydantic** yesterday, I had
to look up the tag format myself to make sure I didn't get it wrong. If
the person who defined the tag format can't remember it, it's probably
best to automate the process.

Also a good chance to argue with @jkeifer about whether we want to try
to do dyanmic versioning in the monorepo 😄

## Checklist

- [x] Tests pass: `uv run pytest`
- [x] Checks pass: `uv run pre-commit --all-files`