refactor method url_for in ProductRouter and RootRouter (#120)
## What I'm changing

- refactor use of Request#url_for into a static method in RootRouter and
ProductRouter

## How I did it

I created a new static method that all of the generated links use.

The intention of this is to allow child class implementations for
projects that use this library to override this if they want to change
the urls. A key use case for this is when the API is being proxied from
another API, and there is the desire for the links to be correct wrt the
URLs provided by the reverse proxy.

## Checklist

- [X] Tests pass: `uv run pytest`
- [X] Checks pass: `uv run pre-commit run --all-files`
- [X] CHANGELOG is updated (if necessary)

---------

Co-authored-by: Phil Varner <phil.varner@orbitalsidekick.com>